changed location of content of ad_spdlog_formatter.h
