+++ cpp/memilio/CMakeLists.txt
@@ -89,7 +89,6 @@ add_library(memilio
     utils/miompi.cpp
     utils/mioomp.h
     utils/type_list.h
-    ad/include/ad/ad_spdlog_formatter.h
 )
 
 target_include_directories(memilio PUBLIC
@@ -99,7 +98,7 @@ target_include_directories(memilio PUBLIC
 )
 
 target_compile_features(memilio PUBLIC cxx_std_17)
-target_link_libraries(memilio PUBLIC spdlog::spdlog Eigen3::Eigen Boost::boost Boost::filesystem Boost::disable_autolinking Random123)
+target_link_libraries(memilio PUBLIC spdlog::spdlog Eigen3::Eigen Boost::boost Boost::filesystem Boost::disable_autolinking Random123 AD::AD)
 target_compile_options(memilio
     PRIVATE
     ${MEMILIO_CXX_FLAGS_ENABLE_WARNING_ERRORS}
